Creating OpenCL implementation of HippoNonbondedForce
